Command Registration and Enhancements
Changed all commands to the proper RegisterCommand, and Began tweaking to have argument based commands for items.

Currently /rw or /doors toggles all windows or doors, but providing an argument will toggle driver window/door.  The goal is to provide 0-4 to toggle the specific window or no args to toggle all.

Trunk and Hood will be rolled into the current doors and the command will be changed to something else.  Perhaps it will be /toggle [trunk,hood,dfd, drd, pfd, prd] or something of that nature.